WindowShapes/requirements.txt
torch
pillow
numpy
matplotlib
opencv-python